Loads and unloads components explicitly
The slave controller now loads and unloads the component on receipt of a load
command from the master. When the master goes offline it commands the slaves
to unload their components but leaves the docker containers running. If the
master is set back online it commands the slaves to re-load the components
in the same containers.

shutdown stops and removes the containers