Adds a timeout function to the state machine
There is now a state machine State class called TimedState which
posts an event if the state hasn't been exited after some interval.

The master controller state machine uses this for the configuring
state and posts an configuring_timeout event after 30s. This event
moves the state machine into the unavailable state.

There is a similar timeout for the uncofiguring state that puts the
statemachine into the standby state.