Sorts out host names and ports inside containers
The view of the network inside a swarm is different from that from
real hosts. Inside the swarm, a container sees the other services as
running on hosts on the docker "sip" network and the ports are the ports
that the services are actually running on. From outside the swarm
everything is accessed via ports on the swarm master that are mapped
by the docker network to the actual containers and ports.

This means that the Docker Paas module has to know whether it is
running inside the swarm or not. I've done this my having the Dockerfile
create a file called "docker_swarm" when it build the container and
DockerPaas changes its behaviour if it finds that the file exists.