Enables restarting the master controller
It is now possible to kill the master controller when it is on-line
and restart it and it re-esablishes connections to the QA and LTS
slaves and automatically go back to the on-line state.

It doesn't work properly if you kill the master controller when it is
in the configuring state because it doesn't know that it has to send
the slave controllers a load command.