Demos running a task with access to the host network
This commit implements a new slave TaskControl class (called
TaskControlIngest) that starts the task in a plain (i.e. not swarm) docker
container with the network_mode set to "host". This gives the task
direct access to the hosts network adapter.

There is no port mapping involved so the container can use any ports it
likes (provided of course that the host isn't using them) so should be
compatible with FTP.

The task called ingest_demo binds to a port passed to it on the command
line (set to 6000 in etc/slave_map). You can use telnet to connect to
this port to demonstrate that it is working.