demonstrates pinning an ingest task to a host
The task_control.start method now has a "service" argument that distinguishes
between services that should be replicated and tasks which shouldn't and
can have a specifi host specified.

The host is defined by a "host" entry in the slave map - the slave map
will need to be edited to specify a host that exists in swarm. If it
doesn't exist or isn't running the swarm service sits in the pending state
waiting for it to become available.