Decoupling the scheduler from sip_pbc
If we want to be able to use other PBC implementations, then the code in
the scheduler should be as PBC-agnostic as possible. By using our own
Celery application instance, and calling tasks purely by name we are
able to break that dependency, and easily call other implementations.

The code used to setup a Celery app is becoming increasingly more
common, so it's probably worth putting it into a centralized function at
some point.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>